Fix dependencies in integration tests (#13321)
This change moves some type/function definitions in integration tests to fix compilation.
