+++ integration/helpers.go
@@ -47,7 +47,6 @@ import (
 	"github.com/gravitational/teleport/api/constants"
 	apidefaults "github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/auth/native"
 	"github.com/gravitational/teleport/lib/auth/testauthority"
@@ -1918,118 +1917,6 @@ func getLocalIP() (string, error) {
 		}
 	}
 	return "", trace.NotFound("No non-loopback local IP address found")
-}
-
-func createTrustedClusterPair(t *testing.T, suite *integrationTestSuite, extraServices func(*testing.T, *TeleInstance, *TeleInstance)) *client.TeleportClient {
-	ctx := context.Background()
-	username := suite.me.Username
-	name := "test"
-	rootName := fmt.Sprintf("root-%s", name)
-	leafName := fmt.Sprintf("leaf-%s", name)
-
-	// Create root and leaf clusters.
-	root := NewInstance(InstanceConfig{
-		ClusterName: rootName,
-		HostID:      HostID,
-		NodeName:    Host,
-		Priv:        suite.priv,
-		Pub:         suite.pub,
-		log:         suite.log,
-		Ports:       standardPortsOrMuxSetup(false),
-	})
-	leaf := NewInstance(InstanceConfig{
-		ClusterName: leafName,
-		HostID:      HostID,
-		NodeName:    Host,
-		Priv:        suite.priv,
-		Pub:         suite.pub,
-		log:         suite.log,
-		Ports:       standardPortsOrMuxSetup(false),
-	})
-
-	role, err := types.NewRole("dev", types.RoleSpecV5{
-		Allow: types.RoleConditions{
-			Logins: []string{username},
-		},
-	})
-	require.NoError(t, err)
-	root.AddUserWithRole(username, role)
-
-	makeConfig := func() (*testing.T, []*InstanceSecrets, *service.Config) {
-		tconf := suite.defaultServiceConfig()
-		tconf.Proxy.DisableWebService = false
-		tconf.Proxy.DisableWebInterface = true
-		tconf.SSH.Enabled = false
-		return t, nil, tconf
-	}
-
-	oldInsecure := lib.IsInsecureDevMode()
-	lib.SetInsecureDevMode(true)
-	defer lib.SetInsecureDevMode(oldInsecure)
-
-	require.NoError(t, root.CreateEx(makeConfig()))
-	require.NoError(t, leaf.CreateEx(makeConfig()))
-	require.NoError(t, leaf.Process.GetAuthServer().UpsertRole(ctx, role))
-
-	// Connect leaf to root.
-	tcToken := "trusted-cluster-token"
-	tokenResource, err := types.NewProvisionToken(tcToken, []types.SystemRole{types.RoleTrustedCluster}, time.Time{})
-	require.NoError(t, err)
-	require.NoError(t, root.Process.GetAuthServer().UpsertToken(ctx, tokenResource))
-	trustedCluster := root.AsTrustedCluster(tcToken, types.RoleMap{
-		{Remote: "dev", Local: []string{"dev"}},
-	})
-
-	require.NoError(t, root.Start())
-	require.NoError(t, leaf.Start())
-
-	t.Cleanup(func() { root.StopAll() })
-	t.Cleanup(func() { leaf.StopAll() })
-
-	require.NoError(t, trustedCluster.CheckAndSetDefaults())
-	tryCreateTrustedCluster(t, leaf.Process.GetAuthServer(), trustedCluster)
-	waitForTunnelConnections(t, root.Process.GetAuthServer(), leafName, 1)
-
-	rootSSHPort := ports.PopInt()
-	rootProxySSHPort := ports.PopInt()
-	rootProxyWebPort := ports.PopInt()
-	require.NoError(t, root.StartNodeAndProxy("root-zero", rootSSHPort, rootProxyWebPort, rootProxySSHPort))
-	leafSSHPort := ports.PopInt()
-	leafProxySSHPort := ports.PopInt()
-	leafProxyWebPort := ports.PopInt()
-	require.NoError(t, leaf.StartNodeAndProxy("leaf-zero", leafSSHPort, leafProxyWebPort, leafProxySSHPort))
-
-	// Add any extra services.
-	if extraServices != nil {
-		extraServices(t, root, leaf)
-	}
-
-	require.Eventually(t, waitForClusters(root.Tunnel, 1), 10*time.Second, 1*time.Second)
-	require.Eventually(t, waitForClusters(leaf.Tunnel, 1), 10*time.Second, 1*time.Second)
-
-	// Create client.
-	creds, err := GenerateUserCreds(UserCredsRequest{
-		Process:        root.Process,
-		Username:       username,
-		RouteToCluster: rootName,
-	})
-	require.NoError(t, err)
-
-	tc, err := root.NewClientWithCreds(ClientConfig{
-		Login:   username,
-		Cluster: rootName,
-		Host:    Loopback,
-		Port:    rootProxySSHPort,
-	}, *creds)
-	require.NoError(t, err)
-
-	leafCAs, err := leaf.Secrets.GetCAs()
-	require.NoError(t, err)
-	for _, leafCA := range leafCAs {
-		require.NoError(t, tc.AddTrustedCA(leafCA))
-	}
-
-	return tc
 }
 
 // disabledIMDSClient is an EC2 instance metadata client that is always disabled. This is faster
